+++ alembic.ini
@@ -3,6 +3,6 @@
 [alembic]
 # path to migration scripts
 # Use forward slashes (/) also on windows to provide an os agnostic path
-script_location = dbos_transact/migrations
+script_location = dbos/migrations
 
 version_path_separator = os  # Use os.pathsep. Default configuration used for new projects.
