+++ dbos/tracer.py
@@ -7,7 +7,7 @@ from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
 from opentelemetry.trace import Span
 
-from dbos_transact.dbos_config import ConfigFile
+from dbos.dbos_config import ConfigFile
 
 if TYPE_CHECKING:
     from .context import TracedAttributes
