+++ tests/test_admin_server.py
@@ -4,7 +4,7 @@ import uuid
 
 import requests
 
-from dbos_transact import DBOS, SetWorkflowUUID
+from dbos import DBOS, SetWorkflowUUID
 
 
 def test_admin_endpoints(dbos: DBOS) -> None:
